+++ src/index.js
@@ -1,6 +1,7 @@
-require('colors')
+require('colors');
 const node = require('./node').node;
 const api = require('./node').api;
+const utils = require('./node').utils;
 
 // Some general TODOs:
 // TODO: add linting
@@ -16,7 +17,7 @@ module.exports = {
 
         _node.start().then((n) => {
             api.createAPI(n);
-            n.log('initialized!'.green.bold);
+            n.log(`Nelson v.${utils.getVersion()} initialized`.green.bold);
         });
     },
     ...node